Add credential handling, account creation + login
Credentials can now be created on /createAccount page. User is then
informed of the success, and given a link to the login page.

If existing username is provided at account creation, server sends
500-response (TODO do it in a nicer way).

Logging in requires that users credentials have been saved in the
credential storage earlier. After successful login user is redirected to
a placeholder web page that holds links to some resources.

If wrong credentials are provided at log in, user is redirected to login
page again (TODO do it in a nicer way).